+++ HISTORY.rst
@@ -26,6 +26,7 @@ latest (unreleased)
 
 * BSSFL-462: Payment Terms on the Internal Purchase Order
 * BSSFL-461: DN/Invoice today date
+* BSSFL-465: Exchange difference move line label
 
 **Bugfixes**